Autenticação CRUD Pedido

+++ SoftCandy/Controllers/PedidoController.cs
@@ -22,69 +22,77 @@ namespace SoftCandy.Controllers
         // GET: Pedido
         public async Task<IActionResult> Index()
         {
-            var softCandyContext = _context.Pedido.Include(c => c.Cliente);
+            if (User.Identity.IsAuthenticated)
+            {
+                var softCandyContext = _context.Pedido.Include(c => c.Cliente);
 
-            return View(await softCandyContext.ToListAsync());
-           
-           
+                return View(await softCandyContext.ToListAsync());
+            }
+            return RedirectToAction("Index", "Home");
+
         }
 
         // GET: Pedido/Details
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
+            if (User.Identity.IsAuthenticated)
             {
-                return NotFound();
-            }
+                if (id == null)
+                {
+                    return NotFound();
+                }
 
-            var pedido = await _context.Pedido
-                .Include(p => p.Cliente)
-                .FirstOrDefaultAsync(m => m.Num_Pedido == id);
-            if (pedido == null)
-            {
-                return NotFound();
-            }
+                var pedido = await _context.Pedido
+                    .Include(p => p.Cliente)
+                    .FirstOrDefaultAsync(m => m.Num_Pedido == id);
+                if (pedido == null)
+                {
+                    return NotFound();
+                }
 
-            return View(pedido);
+                return View(pedido);
+            }
+            return RedirectToAction("Index", "Home");
         }
 
         // GET: Pedido/Create
         public IActionResult Create()
         {
-            ViewData["ID_CLIENTE"] = new SelectList(_context.Cliente, "Id_Cliente", "Nome");
-            return View();
+            if (User.Identity.IsAuthenticated)
+            {
+                ViewData["ID_CLIENTE"] = new SelectList(_context.Cliente, "Id_Cliente", "Nome");
+                return View();
+            }
+            return RedirectToAction("Index", "Home");
         }
 
         // POST: Pedido/Create
         [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Num_Pedido,Valor_Total,Desconto,Data_Pedido,ID_CLIENTE")] Pedido pedido)
+        // [ValidateAntiForgeryToken]
+        public async Task<IActionResult> Create(List<Item_Pedido> itensL)
         {
-            if (ModelState.IsValid)
-            {
-                _context.Add(pedido);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
-            }
-            ViewData["ID_CLIENTE"] = new SelectList(_context.Cliente, "Id_Cliente", "Celular", pedido.ID_CLIENTE);
-            return View(pedido);
+            return View(itensL);
         }
 
         // GET: Pedido/Edit
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
+            if (User.Identity.IsAuthenticated)
             {
-                return NotFound();
-            }
+                if (id == null)
+                {
+                    return NotFound();
+                }
 
-            var pedido = await _context.Pedido.FindAsync(id);
-            if (pedido == null)
-            {
-                return NotFound();
+                var pedido = await _context.Pedido.FindAsync(id);
+                if (pedido == null)
+                {
+                    return NotFound();
+                }
+                ViewData["ID_CLIENTE"] = new SelectList(_context.Cliente, "Id_Cliente", "Celular", pedido.ID_CLIENTE);
+                return View(pedido);
             }
-            ViewData["ID_CLIENTE"] = new SelectList(_context.Cliente, "Id_Cliente", "Celular", pedido.ID_CLIENTE);
-            return View(pedido);
+            return RedirectToAction("Index", "Home");
         }
 
         // POST: Pedido/Edit
@@ -92,52 +100,60 @@ namespace SoftCandy.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("Num_Pedido,Valor_Total,Desconto,Data_Pedido,ID_CLIENTE")] Pedido pedido)
         {
-            if (id != pedido.Num_Pedido)
+            if (User.Identity.IsAuthenticated)
             {
-                return NotFound();
-            }
+                if (id != pedido.Num_Pedido)
+                {
+                    return NotFound();
+                }
 
-            if (ModelState.IsValid)
-            {
-                try
+                if (ModelState.IsValid)
                 {
-                    _context.Update(pedido);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!PedidoExists(pedido.Num_Pedido))
+                    try
                     {
-                        return NotFound();
+                        _context.Update(pedido);
+                        await _context.SaveChangesAsync();
                     }
-                    else
+                    catch (DbUpdateConcurrencyException)
                     {
-                        throw;
+                        if (!PedidoExists(pedido.Num_Pedido))
+                        {
+                            return NotFound();
+                        }
+                        else
+                        {
+                            throw;
+                        }
                     }
+                    return RedirectToAction(nameof(Index));
                 }
-                return RedirectToAction(nameof(Index));
+                ViewData["ID_CLIENTE"] = new SelectList(_context.Cliente, "Id_Cliente", "Celular", pedido.ID_CLIENTE);
+                return View(pedido);
             }
-            ViewData["ID_CLIENTE"] = new SelectList(_context.Cliente, "Id_Cliente", "Celular", pedido.ID_CLIENTE);
-            return View(pedido);
+            return RedirectToAction("Index", "Home");
         }
 
         // GET: Pedido/Delete
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
+            if (User.Identity.IsAuthenticated)
             {
-                return NotFound();
-            }
+                if (id == null)
+                {
+                    return NotFound();
+                }
 
-            var pedido = await _context.Pedido
-                .Include(p => p.Cliente)
-                .FirstOrDefaultAsync(m => m.Num_Pedido == id);
-            if (pedido == null)
-            {
-                return NotFound();
-            }
+                var pedido = await _context.Pedido
+                    .Include(p => p.Cliente)
+                    .FirstOrDefaultAsync(m => m.Num_Pedido == id);
+                if (pedido == null)
+                {
+                    return NotFound();
+                }
 
-            return View(pedido);
+                return View(pedido);
+            }
+            return RedirectToAction("Index", "Home");
         }
 
         // POST: Pedido/Delete
@@ -145,10 +161,14 @@ namespace SoftCandy.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var pedido = await _context.Pedido.FindAsync(id);
-            _context.Pedido.Remove(pedido);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            if (User.Identity.IsAuthenticated)
+            {
+                var pedido = await _context.Pedido.FindAsync(id);
+                _context.Pedido.Remove(pedido);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            return RedirectToAction("Index", "Home");
         }
 
         private bool PedidoExists(int id)
